requirement.txt
mutagen
eyed3
librosa
openai-whisper
requests
numpy
ffmpeg-python
